nixos/etesync-dav: ensure htpaswd file is generated
This commit ensures /var/lib/etesync-dav/htpaswd is generated before starting
the main script.

This file is required by the etesync-dav package and is not a typo; refer to
https://github.com/etesync/etesync-dav/blob/f7ea58328f446732a5fb49742e0d0b5b5155cc47/etesync_dav/config.py#L33C1-L33C50.
Starting etesync-dav for the first time without this file, one will encounter
the "Failed to load htpasswd file /var/lib/etesync-dav exception error.
Creating a blank file manually fixes the issue, but that is unnecessary friction
for users; it is better to amend the ExecStartPre script to create the file if
has yet to exist.